requirements.txt
IPy==0.81
Pillow==2.5.3
Twisted==14.0.0
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
ecdsa==0.11
miniupnpc==1.9
pybitcointools==1.1
pycountry==1.8
pyelliptic==1.5.3
pysqlcipher==2.6.3-1
pystun=0.0.4
python-gnupg==0.3.6
python-obelisk==0.1.3
pyzmq==14.3.1
qrcode==5.0.1
requests==2.4.1
six==1.7.3
tornado==4.0.1
websocket-client==0.16.0
wsgiref==0.1.2
zope.interface==4.1.1
